Update components folders
