+++ aprove-me-next/src/components/assignors/AssignorDialog.tsx
@@ -8,7 +8,6 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog"
-import PayableForm from "./PayableForm"
 import { Edit } from "lucide-react";
 import { useState } from "react";
 import AssignorForm from "./AssignorForm";
