+++ aprove-me-next/src/components/payables/PayablesTable.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useEffect, useState } from 'react'
-import { Table, TableBody, TableCaption, TableCell, TableHead, TableHeader, TableRow } from './ui/table'
+import { Table, TableBody, TableCaption, TableCell, TableHead, TableHeader, TableRow } from '../ui/table'
 import { usePayables } from '@/hooks/usePayables';
 import { formatDateBR } from '@/lib/formatDate';
 import Link from 'next/link';
